When not preserving 'due', ensure it's scheduled as a new card
I've been running into an issue where I copy from existing cards
I've already reviewed before, but the due field gets copied over
from the already reviewed card, creating a 'due' which is
very low and confuses Anki into scheduling the card as a
non-new card, even though it displays as a new card in the
browser. For example, it'll say something like "New #1234"
in the browser instead of something like "New #2000001",
where that 1234 is the due date in days of the old card,
and something in Anki sees that and schedules it like
I've already reviewed it for some reason.

Because of this, I have to select my copied cards and click
"Reposition" or "Forget" to make the due field update
appropriately. This fix is to make that happen automatically.