Ensure new cards are scheduled as new
When using the V2 scheduler, there's a field called 'queue' on the cards table that should be set to 0 when copying to a new card, but it's not getting reset when ignoring scheduling information. This creates a weird situation where the new card gets displayed as "New" in the Anki browser but won't necessarily appear in your queue of new cards, specifically if the card you copied has already been reviewed before. The best way to guarantee this doesn't happen should be to use the same function Anki uses when you click "Forget" or "Reposition" on a card.